# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fazenda São Bento — Sistema de Gestão Agrícola
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
...
  /trucks:
    get:
      operationId: listTrucks
      tags: [trucks]
      summary: List all trucks
      responses:
        "200":
          description: List of trucks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Truck"
    post:
      operationId: createTruck
      tags: [trucks]
      summary: Create truck
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTruckRequest"
      responses:
        "201":
          description: Truck created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Truck"
  /trucks/{id}:
    put:
      operationId: updateTruck
      tags: [trucks]
      summary: Update truck
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTruckRequest"
      responses:
        "200":
          description: Truck updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Truck"
    delete:
      operationId: deleteTruck
      tags: [trucks]
      summary: Delete truck
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Truck deleted
...
  /transport:
    get:
      operationId: listTransport
      tags: [transport]
      summary: List transport records
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of transport records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransportRecord"
    post:
      operationId: createTransport
      tags: [transport]
      summary: Create transport record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransportRequest"
      responses:
        "201":
          description: Transport record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransportRecord"
  /transport/{id}:
    put:
      operationId: updateTransport
      tags: [transport]
      summary: Update transport record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransportRequest"
      responses:
        "200":
          description: Transport record updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransportRecord"
    delete:
      operationId: deleteTransport
      tags: [transport]
      summary: Delete transport record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transport record deleted
...
  /stock-movements:
    get:
      operationId: listStockMovements
      tags: [stock-movements]
      summary: List stock movements
      parameters:
        - name: productId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [entrada, saida]
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of stock movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockMovement"
    post:
      operationId: createStockMovement
      tags: [stock-movements]
      summary: Create stock movement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStockMovementRequest"
      responses:
        "201":
          description: Stock movement created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockMovement"
        "400":
          description: Insufficient stock
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stock-movements/{id}:
    delete:
      operationId: deleteStockMovement
      tags: [stock-movements]
      summary: Delete stock movement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Movement deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
      required:
        - message
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, operador]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - createdAt
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [admin, operador]
      required:
        - name
        - email
        - password
        - role
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, operador]
        password:
          type: string
...
    Truck:
      type: object
      properties:
        id:
          type: integer
        plate:
          type: string
        model:
          type: string
        capacity:
          type: number
        status:
          type: string
          enum: [ativo, manutencao, inativo]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - plate
        - status
        - createdAt
    CreateTruckRequest:
      type: object
      properties:
        plate:
          type: string
        model:
          type: string
        capacity:
          type: number
        status:
          type: string
          enum: [ativo, manutencao, inativo]
      required:
        - plate
        - status
...
    TransportRecord:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        truckId:
          type: integer
        truckPlate:
          type: string
        driverName:
          type: string
        origin:
          type: string
        destination:
          type: string
        cargoTons:
          type: number
        freightValue:
          type: number
        harvestId:
          type: integer
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - date
        - truckId
        - driverName
        - origin
        - destination
        - cargoTons
        - createdAt
    CreateTransportRequest:
      type: object
      properties:
        date:
          type: string
          format: date
        truckId:
          type: integer
        driverName:
          type: string
        origin:
          type: string
        destination:
          type: string
        cargoTons:
          type: number
        freightValue:
          type: number
        harvestId:
          type: integer
        notes:
          type: string
      required:
        - date
        - truckId
        - driverName
        - origin
        - destination
        - cargoTons
...
    StockMovement:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        type:
          type: string
          enum: [entrada, saida]
        quantity:
          type: number
        reason:
          type: string
        culture:
          type: string
        operatorName:
          type: string
        date:
          type: string
          format: date
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - productId
        - type
        - quantity
        - date
        - createdAt
    CreateStockMovementRequest:
      type: object
      properties:
        productId:
          type: integer
        type:
          type: string
          enum: [entrada, saida]
        quantity:
          type: number
        reason:
          type: string
        culture:
          type: string
        operatorName:
          type: string
        date:
          type: string
          format: date
        notes:
          type: string
      required:
        - productId
        - type
        - quantity
        - date
...