# lib/api-zod/src/generated/types/stockMovement.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fazenda São Bento — Sistema de Gestão Agrícola
 * OpenAPI spec version: 0.1.0
 */
import type { StockMovementType } from "./stockMovementType";

export interface StockMovement {
  id: number;
  productId: number;
  productName?: string;
  type: StockMovementType;
  quantity: number;
  reason?: string;
  culture?: string;
  operatorName?: string;
  date: Date;
  notes?: string;
  createdAt: Date;
}